chore: p0-group-detail-crash-fix-001セッション終了 - P0クラッシュ修正完了（81点）
## 完了タスク

### 1. P0グループ詳細クラッシュ修正
- PhotoThumbnail.swift: Continuation二重resume問題解消
  - withCheckedContinuationパターン採用
  - onDisappearでPHImageRequestキャンセル
  - Task.isCancelledチェック3箇所追加
- GroupDetailView.swift: エラーハンドリング強化
  - 空グループチェック追加
  - 非同期処理キャンセルチェック

### 2. テスト生成（54件）
- PhotoThumbnailTests.swift: 24テストケース
- GroupDetailViewTests.swift: 30テストケース
- TestTags.swift: 共通Tag定義（重複エラー解消）

### 3. E2Eテスト実行
- シミュレーター: 94.4%合格（17/18）
- 実機テスト: 準備完了（検証待ち）

### 4. コード整理
- 不要なテストファイル削除（4件）
- SESSION_SUMMARYをアーカイブへ移動

## 品質スコア: 81点（条件付き合格）

## 次回タスク
- 実機テスト結果確認
- 品質スコア90点達成
- M10-T04〜T06（リリース準備）

+++ LightRoll_CleanerPackage/Tests/LightRoll_CleanerFeatureTests/ImageAnalysis/ConcurrencyPerformanceTests.swift
@@ -173,7 +173,7 @@ struct PerformanceTests {
         }
 
         let endTime = ContinuousClock.now
-        return startTime.duration(to: endTime).components.attoseconds / 1_000_000_000_000_000
+        return Int(startTime.duration(to: endTime).components.attoseconds / 1_000_000_000_000_000)
     }
 
     /// 並列実行をシミュレート
@@ -189,7 +189,7 @@ struct PerformanceTests {
         }
 
         let endTime = ContinuousClock.now
-        return startTime.duration(to: endTime).components.attoseconds / 1_000_000_000_000_000
+        return Int(startTime.duration(to: endTime).components.attoseconds / 1_000_000_000_000_000)
     }
 
     // MARK: - テスト1: 直列vs並列の速度比較
